app/career_page/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain_huggingface
python-dotenv
pydantic
requests
